crypto/ethereumbook/requirements.txt
web3
py-solc-x
eth-utils
